Remove the security system JWT Refactor the class Client to User Adding the process of validation from username by Header

+++ Back/src/main/java/com/br/decea/ServiceImpl/ProjectService.java
@@ -5,15 +5,16 @@ import com.br.decea.DTO.ProjectResponseDTO;
 import com.br.decea.DTO.ProjectUpdateDTO;
 import com.br.decea.External.Comunication;
 import com.br.decea.External.Location;
-import com.br.decea.Model.Client;
+import com.br.decea.Model.User;
 import com.br.decea.Model.Project;
-import com.br.decea.Repository.ClientRepository;
 import com.br.decea.Repository.ProjectRepository;
 import com.br.decea.Service.IProjectService;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.br.decea.Repository.UserRepository;
+import com.br.decea.Service.IUserService;
 
 /**
  *
@@ -26,32 +27,39 @@ public class ProjectService implements IProjectService {
     private ProjectRepository projectRepository;
 
     @Autowired
-    private ClientRepository clientRepository;
+    private IUserService userService;
 
     @Override
-    public Project create(ProjectDTO projectDTO) {
+    public Project create(ProjectDTO projectDTO) throws Exception {
 
-        Project auxProject = new Project();
+        try {
+            User user = userService.getUser(projectDTO.getUsername());
 
-        auxProject.setTitle(projectDTO.getTitle());
-        auxProject.setZip_code(projectDTO.getZip_code());
-        auxProject.setCost(projectDTO.getCost());
-        auxProject.setDone(projectDTO.getDone());
+            Project auxProject = new Project();
+            auxProject.setUser(user);
+            auxProject.setTitle(projectDTO.getTitle());
+            auxProject.setZip_code(projectDTO.getZip_code());
+            auxProject.setCost(projectDTO.getCost());
+            auxProject.setDone(projectDTO.getDone());
 
-        Optional<Client> optClient = clientRepository.findByUsername(projectDTO.getUsername());
-        auxProject.setClient(optClient.get());
+            auxProject.setDeadline(projectDTO.getDeadline());
+            auxProject.setCreated_at(projectDTO.getCreated_at());
+            auxProject.setUpdated_at(projectDTO.getUpdated_at());
 
-        auxProject.setDeadline(projectDTO.getDeadline());
-        auxProject.setCreated_at(projectDTO.getCreated_at());
-        auxProject.setUpdated_at(projectDTO.getUpdated_at());
-
-        return projectRepository.save(auxProject);
-
+            return projectRepository.save(auxProject);
+        } catch (Exception e) {
+            throw new Exception("User not found");
+        }
     }
 
     @Override
-    public List<Project> getAll() {
-        return projectRepository.findAll();
+    public List<Project> getAll(String username) throws Exception {
+        try {
+            User user = userService.getUser(username);
+            return user.getProjects();
+        } catch (Exception e) {
+            throw new Exception("User not found");
+        }
     }
 
     @Override
@@ -70,7 +78,7 @@ public class ProjectService implements IProjectService {
         projectResponseDTO.setTitle(auxProject.getTitle());
         projectResponseDTO.setCost(auxProject.getCost());
         projectResponseDTO.setDone(auxProject.getDone());
-        projectResponseDTO.setUsername(auxProject.getClient().getUsername());
+        projectResponseDTO.setUsername(auxProject.getUser().getUsername());
         projectResponseDTO.setDeadline(auxProject.getDeadline());
         projectResponseDTO.setCreated_at(auxProject.getCreated_at());
         projectResponseDTO.setUpdated_at(auxProject.getUpdated_at());
@@ -79,12 +87,16 @@ public class ProjectService implements IProjectService {
     }
 
     @Override
-    public Project update(Long id, ProjectUpdateDTO projectDTO) throws Exception {
+    public Project update(Long id, ProjectUpdateDTO projectDTO, String username) throws Exception {
         Project auxProject = projectRepository.findById(id)
                 .orElseThrow(() -> new Exception("Project not found"));
 
         if (auxProject.getDone()) {
             throw new Exception("Project had done");
+        }
+
+        if (auxProject.getUser().getUsername() != username) {
+            throw new Exception("The user can't do any alteration");
         }
 
         auxProject.setTitle(projectDTO.getTitle());
@@ -95,20 +107,29 @@ public class ProjectService implements IProjectService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id,String username) throws Exception {
+        Project project = projectRepository.findById(id).orElseThrow(
+        ()-> new Exception("Project not found"));
+        
+        if (project.getUser().getUsername() != username) {
+            throw new Exception("The user can't do any alteration");
+        }
+        
         projectRepository.deleteById(id);
     }
 
     @Override
-    public void doneProject(Long id) throws Exception {
-        Optional<Project> optProject = projectRepository.findById(id);
+    public void doneProject(Long id, String username) throws Exception {
 
-        if (optProject.isEmpty()) {
-            throw new Exception("Project not found");
+        Project project = projectRepository.findById(id).orElseThrow(
+                () -> new Exception("Project not found"));
+
+        if (project.getUser().getUsername() != username) {
+            throw new Exception("The user can't do any alteration");
         }
 
-        optProject.get().setDone(Boolean.TRUE);
-        projectRepository.save(optProject.get());
+        project.setDone(Boolean.TRUE);
+        projectRepository.save(project);
     }
 
 }
